requirements.txt
blinker==1.6.2
boto3==1.26.136
botocore==1.29.136
click==8.1.3
Flask==2.3.2
Flask-Cors==3.0.10
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.2
psycopg2==2.9.6
psycopg2-binary==2.9.6
python-dateutil==2.8.2
python-dotenv==1.0.0
s3transfer==0.6.1
six==1.16.0
urllib3==1.26.15
Werkzeug==2.3.4
